requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
chromadb
langchain
langchain-community
langchain-chroma
langchain-openai
pdfplumber
pandas
openpyxl
requests
